Multiple math::matrix-related changes
- Added math compile-time configuration.
- Optimized math::fast_multiply.
- Minor updates in benchmarks.

+++ benchmark/math_vector_hsum-4i32.cpp
@@ -4,7 +4,7 @@
  *  LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
  */
 #include <cstdint>
-typedef int64_t T;
+typedef int32_t T;
 #define EAGINE_BM_N 4
 #include "math_vector_hsum.ipp"
